1.命名更改 2.改int为long

+++ src/main/java/nju/dao/AppointmentDao.java
@@ -3,6 +3,6 @@ package nju.dao;
 import nju.model.AppointmentData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AppointmentDao extends JpaRepository<AppointmentData, Integer> {
+public interface AppointmentDao extends JpaRepository<AppointmentData, Long> {
 
 }
