+++ src/main/java/nju/dao/RecordDao.java
@@ -3,6 +3,6 @@ package nju.dao;
 import nju.model.RecordData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecordDao extends JpaRepository<RecordData, Integer> {
+public interface RecordDao extends JpaRepository<RecordData, Long> {
 
 }
