+++ src/main/java/nju/dao/StaffDao.java
@@ -3,6 +3,6 @@ package nju.dao;
 import nju.model.StaffData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StaffDao extends JpaRepository<StaffData, Integer> {
+public interface StaffDao extends JpaRepository<StaffData, Long> {
 
 }
